added receptionist screen

+++ frontend/app/src/main/java/com/fmi_unitbv2026/kronsoft_frontend/data/models/CreatePatient.kt
@@ -24,5 +24,5 @@ data class CreatePatient(
     val cnp: String,
 
     @SerializedName("idDoctor")
-    val idDoctor: Int
+    val idDoctor: Long
 )
